+++ uspace/srv/fs/exfat/exfat_ops.c
@@ -88,7 +88,7 @@ static unsigned exfat_lnkcnt_get(fs_node_t *);
 static bool exfat_is_directory(fs_node_t *);
 static bool exfat_is_file(fs_node_t *node);
 static service_id_t exfat_service_get(fs_node_t *node);
-static long exfat_size_block(service_id_t);
+static uint32_t exfat_size_block(service_id_t);
 
 /*
  * Helper functions.
@@ -912,7 +912,7 @@ service_id_t exfat_service_get(fs_node_t *node)
 	return 0;
 }
 
-long exfat_size_block(service_id_t service_id)
+uint32_t exfat_size_block(service_id_t service_id)
 {
 	exfat_bs_t *bs;
 	bs = block_bb_get(service_id);
